+++ test/scope_test.dart
@@ -1118,6 +1118,178 @@ void main() {
       });
     });
 
+    group('plugins', () {
+      group('addPlugin(plugin)', () {
+        group('should throw', () {
+          test('if the plugin is already added', () {
+            const plugin = ScopePlugin(nodePlugins: {});
+            scope.addPlugin(plugin);
+            expect(
+              () => scope.addPlugin(plugin),
+              throwsA(
+                predicate<ArgumentError>(
+                  (e) => e.toString().contains(
+                        'Plugin already added.',
+                      ),
+                ),
+              ),
+            );
+          });
+
+          test('when there are multiple node plugins with the same key', () {
+            final plugin0 = ScopePlugin(
+              nodePlugins: {
+                'node0': NodeBluePrint.example(key: 'key'),
+                'node1': NodeBluePrint.example(key: 'key'),
+              },
+            );
+
+            expect(
+              () => scope.addPlugin(plugin0),
+              throwsA(
+                predicate<ArgumentError>(
+                  (e) => e.toString().contains(
+                        'Found multiple node plugins with key "key"',
+                      ),
+                ),
+              ),
+            );
+          });
+          test('when host nodes cannot be found', () {
+            final plugin = ScopePlugin(
+              nodePlugins: {
+                'unknown0': NodeBluePrint.example(key: 'unknown'),
+                'unknown1': NodeBluePrint.example(key: 'unknown1'),
+              },
+            );
+            expect(
+              () => scope.addPlugin(plugin),
+              throwsA(
+                predicate<ArgumentError>(
+                  (e) => e.toString().contains(
+                        'Host nodes not found: unknown0, unknown1',
+                      ),
+                ),
+              ),
+            );
+          });
+
+          test('when types of host nodes and plugin nodes do not match', () {
+            final scope = Scope.example();
+            scope.mockContent({
+              'node0': 'string',
+            });
+
+            const plugin = ScopePlugin(
+              nodePlugins: {
+                'node0': NodeBluePrint<int>(key: 'plugin0', initialProduct: 0),
+              },
+            );
+
+            expect(
+              () => scope.addPlugin(plugin),
+              throwsA(
+                predicate<Error>(
+                  (e) => e.toString().contains(
+                        '\'NodeBluePrint<int>\' is not a subtype '
+                        'of type \'NodeBluePrint<String>\'',
+                      ),
+                ),
+              ),
+            );
+          });
+        });
+
+        test('should add the plugin to the list of plugins', () {
+          const plugin = ScopePlugin(nodePlugins: {});
+          scope.addPlugin(plugin);
+          expect(scope.plugins, [plugin]);
+        });
+
+        test('should add the node plugins to their corresponding hosts', () {
+          // Define an existing node and scope hierarchy
+          final scope = Scope.example();
+          scope.mockContent({
+            'a': {
+              'b': {
+                'n0': 0,
+              },
+              'c': {
+                'n1': 0,
+              },
+            },
+          });
+
+          // Get the two nodes we want to add plugins to
+          final node0 = scope.findNode<int>('a.b.n0')!;
+          final node1 = scope.findNode<int>('a.c.n1')!;
+
+          // Define a node plugin that modifies the two nodes
+          final scopePlugin = ScopePlugin(
+            nodePlugins: {
+              'b.n0': NodeBluePrint<int>(
+                key: 'byTwo',
+                initialProduct: 0,
+                produce: (components, previousProduct) =>
+                    (components.first as int) * 2,
+              ),
+              'a.c.n1': NodeBluePrint<int>(
+                key: 'byThree',
+                initialProduct: 0,
+                produce: (components, previousProduct) =>
+                    (components.first as int) * 3,
+              ),
+            },
+          );
+
+          // Add the plugin to the scope
+          scope.addPlugin(scopePlugin);
+
+          // The plugins should have been added to the nodes
+          expect(node0.plugins, hasLength(1));
+          expect(node0.plugins.first.key, 'byTwo');
+          expect(node1.plugins, hasLength(1));
+          expect(node1.plugins.first.key, 'byThree');
+
+          // Remove the plugin from the scope
+          scope.removePlugin(scopePlugin);
+
+          // The plugins should have been removed from the nodes
+          expect(node0.plugins, isEmpty);
+          expect(node1.plugins, isEmpty);
+        });
+      });
+
+      group('removePlugin(plugin)', () {
+        group('should throw', () {
+          test('if the plugin is not added', () {
+            const plugin = ScopePlugin(nodePlugins: {});
+            expect(
+              () => scope.removePlugin(plugin),
+              throwsA(
+                predicate<ArgumentError>(
+                  (e) => e.toString().contains(
+                        'Plugin not found.',
+                      ),
+                ),
+              ),
+            );
+          });
+        });
+
+        test('should remove the plugin from the list of plugins', () {
+          const plugin = ScopePlugin(nodePlugins: {});
+          scope.addPlugin(plugin);
+          scope.removePlugin(plugin);
+          expect(scope.plugins, isEmpty);
+        });
+
+        test('should remove the node plugins from it\'s hosts', () {
+          // Is tested in addPlugin()
+        });
+      });
+    });
+
     group('mockContent', () {
       test('should create a mock content', () {
         final scope = Scope.example();